+++ rtengine/dcrop.h
@@ -18,11 +18,7 @@
  */
 #pragma once
 
-#include "improccoordinator.h"
 #include "rtengine.h"
-#include "improcfun.h"
-#include "imagesource.h"
-#include "procevents.h"
 #include "pipettebuffer.h"
 #include "../rtgui/threadutils.h"
 
@@ -30,6 +26,7 @@ namespace rtengine
 {
 
 class Image8;
+class CieImage;
 
 using namespace procparams;
